:recycle: minor refactor
