+++ data/src/main/java/com/free/data/datasources/GithubUsersPagingSource.kt
@@ -28,7 +28,7 @@ class GithubUsersPagingSource(
                     is Result.Error -> LoadResult.Error(result.exception)
                     is Result.Success -> {
                         LoadResult.Page(
-                            data = result.data.children.map { model -> model.entity },
+                            data = result.data.children,
                             prevKey = result.data.params.since,
                             nextKey = result.data.children.lastOrNull()?.id
                         )
