+++ data/src/main/java/com/free/data/datasources/MockGithubApi.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.free.core.Result
 import com.free.data.models.UserDetailModel
 import com.free.data.models.UserModel
+import com.free.domain.entities.UserDetail
 import com.free.domain.usecases.FetchUsersInputParams
 import com.free.domain.usecases.GetUserDetailInputParams
 import dagger.hilt.android.qualifiers.ApplicationContext
@@ -30,11 +31,11 @@ class MockGithubApi @Inject constructor(
         // ストレージアクセスのラグ
         delay(2000)
         return Result.Success(
-            ListingData(children, params)
+            ListingData(children.map { model -> model.entity }, params)
         )
     }
 
-    override suspend fun userDetail(params: GetUserDetailInputParams): Result<UserDetailModel> {
+    override suspend fun userDetail(params: GetUserDetailInputParams): Result<UserDetail> {
         return Result.Success(
             UserDetailModel(
                 id = 0,
@@ -61,7 +62,7 @@ class MockGithubApi @Inject constructor(
                 following = 1234567890,
                 updatedAt = "2022-04-03T23:40:06Z",
                 createdAt = "2007-10-20T05:24:19Z"
-            )
+            ).entity
         )
     }
 }
