+++ presentation/src/test/java/com/free/presentation/viewmodels/GithubUsersViewModelTest.kt
@@ -1,0 +1,10 @@
+package com.free.presentation.viewmodels
+
+import org.junit.Test
+
+class GithubUsersViewModelTest {
+    @Test
+    fun failureUsersByNetworkError() {
+
+    }
+}
